pyproject.toml
[project]
name = "learning-mcp"
version = "0.1.0"
description = "Learning MCP"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles==24.1.0",
    "black==25.12.0",
    "dotenv==0.9.9",
    "fastmcp==2.14.0",
    "grpcio-tools==1.73.1",
    "langchain>=1.2.6",
    "langchain-google-genai==4.0.0",
    "langchain-mcp-adapters==0.2.1",
    "langchain-ollama==1.0.1",
    "langchain-openai==1.1.3",
    "mcp>=1.24.0",
    "pillow>=10.4.0",
    "speechrecognition==3.14.1",
    "authlib>=1.6.5,<2",
    "pyaudio>=0.2.14",
    "audioop-lts>=0.2.1",
]

[tool.black]
target-version = ["py313", "py314"]
line-length = 100
extend-exclude = "proto_gen*"
